Add basic API server code
This branch is for testing how Dockerizing an API, deploying it, and having multiple instances of it running would work.

No tests or anything of the sort will be added, as the code will be changed many times, and it has no need to be robust, it's just a test of something that may look similar to the actual Gateway server in the future.