comment out handler tests
Architecture is likely to change, as there's lots to learn about
entities used to represent requests vs the entities that are to be saved
in the database. There's also lots to learn about using an OAS spec and
ensuring that it's in sync with the actual API code, as well as actually
validating and testing the functionality of the spec, meaning,
validating that certain requests, given certain state, result in
specific responses...Maybe WireMock, contract testing, or just good ol'
integration testng is the play here...whatever it is, it should be done
later as there's still a lot unknown about the architecture of the
project!